Add instructions for using when embedded in Spring Boot
This plugin gets extensively used when Camunda
is embedded into the Spring Boot application.

Although the project has existed for quite some time,
new users are still struggling to use it with Spring Boot
since searching forums and GitHub issues is not
the optimal way.

This commit adds instructions for new Spring Boot users
so that they can quickly use the plugin, addressing
GitHub issues #30, #24, #15, #8, and #7.